+++ src/AngleSharp.Css/Dom/TextOverflow.cs
@@ -1,0 +1,17 @@
+namespace AngleSharp.Css.Dom
+{
+	/// <summary>
+	/// An enumeration with all possible Text Overflow options.
+	/// </summary>
+	public enum TextOverflow : byte
+	{
+		/// <summary>
+		/// Indicates that clipped content is fully clipped, with no readable indication.
+		/// </summary>
+		Clip,
+		/// <summary>
+		/// Indicates that clipped content is represented by an ellipsis ("…").
+		/// </summary>
+		Ellipsis
+	}
+}